+++ src/main/java/com/easynull/hemomancy/core/compact/curios/CuriosStorage.java
@@ -1,0 +1,4 @@
+//package com.easynull.hemomancy.core.compact.curios;
+//
+//public final class CuriosStorage {
+//}