+++ src/main/java/com/easynull/hemomancy/core/mixin/GameModeMixin.java
@@ -1,0 +1,33 @@
+package com.easynull.hemomancy.core.mixin;
+
+import com.easynull.hemomancy.registers.items.armortools.Desecrated;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.server.level.ServerPlayerGameMode;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.SwordItem;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(ServerPlayerGameMode.class)
+public final class GameModeMixin {
+    @Shadow
+    @Final
+    protected ServerPlayer player;
+
+    @Inject(method = "destroyBlock", at = @At("HEAD"))
+    private void hc$destroyBlock(BlockPos pos, CallbackInfoReturnable<Boolean> cir) {
+        ItemStack stack = player.getMainHandItem();
+        if (player.isSilent()) return;
+        if (!(stack.getItem() instanceof Desecrated)) return;
+        Level level = player.level();
+        BlockState state = level.getBlockState(pos);
+        stack.mineBlock(level, state, pos, player);
+    }
+}