+++ src/main/java/com/easynull/hemomancy/registers/HcBlockEntities.java
@@ -13,4 +13,5 @@ public final class HcBlockEntities {
 
     public static final Supplier<BlockEntityType<AltarBE>> bloodAltar = types.register("blood_altar", ()-> new BlockEntityType<>(AltarBE::new, HcElements.bloodAltar.get()));
     public static final Supplier<BlockEntityType<AlchemyTableBE>> alchemyTable = types.register("alchemy_table", ()-> new BlockEntityType<>(AlchemyTableBE::new, HcElements.alchemyTable.get()));
+//    public static final Supplier<BlockEntityType<RelayBE>> arterialRelay = types.register("arterial_relay", ()-> new BlockEntityType<>(RelayBE::new, HcElements.relay.get()));
 }