+++ src/main/java/com/easynull/hemomancy/registers/items/armortools/Desecrated.java
@@ -1,0 +1,4 @@
+package com.easynull.hemomancy.registers.items.armortools;
+
+public interface Desecrated {
+}